pyproject.toml
[project]
name = "device-data-service"
version = "0.1.0"
packages = [{ include = "src" }]
description = ""
authors = [
    {name = "nishantmpathak",email = "nishantpathak97@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "paho-mqtt (>=2.1.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false